Issue #2397009 by thehong: Rename truefalse module to quizz_truefalse.
