+++ modules/question_types/quiz_page/quiz_page.module
@@ -1,21 +1,6 @@
 <?php
 
 /**
- * @file quiz_page.module
- *
- * Pages are implemented as a question type. They do not affect the score.
- */
-
-/**
- * Implements hook_help().
- */
-function quiz_page_help($path, $args) {
-  if ($path == 'admin/help#quiz_pages') {
-    return t('Provides pages which can contain questions in a quiz.');
-  }
-}
-
-/**
  * Implements hook_quiz_question_info().
  */
 function quiz_page_quiz_question_info() {
@@ -25,7 +10,7 @@ function quiz_page_quiz_question_info() {
           'description'       => t('Quiz pages allow you display questions across multiple pages.'),
           'question provider' => 'QuizPagesQuestion',
           'response provider' => 'QuizPagesResponse',
-          'module'            => 'zquiz_question', // All wrapper functions are in that module.
+          'module'            => 'quizz_question', // All wrapper functions are in that module.
       ),
   );
 }
