+++ modules/question_types/quizz_truefalse/README.md
@@ -1,0 +1,4 @@
+True/False question type for quizz module
+====
+
+Allows the creation of "True or False" questions.
