+++ modules/question_types/quizz_truefalse/quizz_truefalse.module
@@ -1,0 +1,48 @@
+<?php
+
+/**
+ * Implements hook_quiz_question_info().
+ */
+function quizz_truefalse_quiz_question_info() {
+  return array(
+      'truefalse' => array(
+          'name'              => t('True/false question'),
+          'description'       => t('Quiz questions that allow a user to select "true" or "false" as his response to a statement.'),
+          'question provider' => 'Drupal\quizz_truefalse\TrueFalseQuestion',
+          'response provider' => 'Drupal\quizz_truefalse\TrueFalseResponse',
+          'module'            => 'quizz_question', // All wrapper functions are in that module.
+      ),
+  );
+}
+
+/**
+ * Implements hook_theme().
+ */
+function quizz_truefalse_theme() {
+  return array(
+      'truefalse_answering_form' => array(
+          'render element' => 'form',
+          'path'           => drupal_get_path('module', 'truefalse') . '/theme',
+          'template'       => 'truefalse-answering-form',
+      ),
+      'truefalse_response'       => array(
+          'variables' => array(
+              'metadata' => array(),
+              'data'     => array(),
+          ),
+      ),
+  );
+}
+
+/**
+ * Theme the response part of the response report
+ *
+ * @param array $metadata Can be used as a table header
+ * @param array $data Can be used as table rows
+ */
+function theme_truefalse_response($variables) {
+  return theme('table', array(
+      'header' => $variables['metadata'],
+      'rows'   => $variables['data']
+  ));
+}
