+++ modules/question_types/quizz_truefalse/theme/truefalse-answering-form.tpl.php
