+++ tests/QuizAnswerTestCase.test
@@ -2,6 +2,9 @@
 
 class QuizAnswerTestCase extends QuizTestCase {
 
+  protected $extraModules = array('quizz_truefalse', 'text', 'field_ui');
+  protected $questionHandlers = array('truefalse');
+
   public static function getInfo() {
     return array(
         'name'        => 'Quiz answer',
@@ -11,9 +14,6 @@ class QuizAnswerTestCase extends QuizTestCase {
   }
 
   public function setUp($modules = array(), $admin_permissions = array(), $user_permissions = array()) {
-    $modules[] = $this->questionHandler = 'truefalse';
-    $modules[] = 'text';
-    $modules[] = 'field_ui';
     $admin_permissions[] = 'administer quiz questions';
     parent::setUp($modules, $admin_permissions, $user_permissions);
   }
