+++ tests/QuizUpgradeTestCase.test
@@ -7,7 +7,7 @@ abstract class QuizUpgradeBaseTestCase extends UpdatePathTestCase {
   protected static $testDumpFile = '…';
   protected static $dependencies = array('ctools', 'entity', 'filter', 'views', 'views_bulk_operations', 'xautoload');
   protected static $testDescription = 'Test an upgrade from various Quiz versions.';
-  protected static $subModules = array('long_answer', 'matching', 'multichoice', 'quiz_ddlines', 'quiz_directions', 'quiz_page', 'quizz_question', 'quizz_scale', 'short_answer', 'truefalse');
+  protected static $subModules = array('long_answer', 'matching', 'multichoice', 'quiz_ddlines', 'quiz_directions', 'quiz_page', 'quizz_question', 'quizz_scale', 'short_answer', 'quizz_truefalse');
 
   /**
    * Some warnings are produced by other modules, we cann't do anything, just
@@ -35,6 +35,7 @@ abstract class QuizUpgradeBaseTestCase extends UpdatePathTestCase {
     db_query($sql_1, array(':module' => 'module', 'old_name' => 'quiz', 'name' => 'quizz'));
     db_query($sql_1, array(':module' => 'module', 'old_name' => 'quiz_question', 'name' => 'quizz_question'));
     db_query($sql_1, array(':module' => 'module', 'old_name' => 'scale', 'name' => 'quizz_scale'));
+    db_query($sql_1, array(':module' => 'module', 'old_name' => 'truefalse', 'name' => 'quizz_truefalse'));
 
     module_enable(static::$dependencies);
     registry_rebuild();
